+++ debezium-core/src/main/java/io/debezium/pipeline/metrics/traits/ConnectionMetricsMXBean.java
@@ -13,5 +13,5 @@ public interface ConnectionMetricsMXBean {
     boolean isConnected();
 
     /** Numeric companion to {@link #isConnected()}: 1 when connected, 0 otherwise. */
-    long getConnectedCode();
+    int getConnectedCode();
 }
