+++ debezium-core/src/main/java/io/debezium/relational/history/SchemaHistoryMXBean.java
@@ -23,7 +23,7 @@ public interface SchemaHistoryMXBean {
     String getStatus();
 
     /** Numeric companion to {@link #getStatus()}: 0=STOPPED, 1=RECOVERING, 2=RUNNING. */
-    long getStatusCode();
+    int getStatusCode();
 
     /**
      * @return time in epoch seconds when recovery has started
